Revert "Merge branch 'master' into (optimize)-_splitAlcehmica"
This reverts commit 3f8b894362ae383271f89d5c8bdefd892c32d972, reversing
changes made to 2eb80ed73f36b41aea30357dae50aacba16a5d54.
