+++ .gitignore
@@ -19,4 +19,3 @@ package.lock.json
 out/*
 data/allInstallations.json
 data/allTiles.json
-scripts/team/*
