+++ contracts/InstallationDiamond/facets/TestInstallationFacet.sol
@@ -22,41 +22,78 @@ contract TestInstallationFacet is Modifiers {
   );
 
   event UpgradeFinalized(uint256 indexed _realmId, uint256 _coordinateX, uint256 _coordinateY, uint256 _newInstallationId);
+
   event UpgradeQueued(address indexed _owner, uint256 indexed _realmId, uint256 indexed _queueIndex);
   event UpgradeQueueFinalized(address indexed _owner, uint256 indexed _realmId, uint256 indexed _queueIndex);
-  event UpgradeTimeReduced(uint256 indexed _queueId, uint256 indexed _realmId, uint256 _coordinateX, uint256 _coordinateY, uint40 _blocksReduced);
 
-  function mockUpgradeInstallation(
-    UpgradeQueue memory _upgradeQueue,
-    uint256 _gotchiId,
-    uint40 _gltr
-  ) external {
-    // Storing variables in memory needed for validation and execution
-    uint256 nextLevelId = s.installationTypes[_upgradeQueue.installationId].nextLevelId;
-    InstallationType memory nextInstallation = s.installationTypes[nextLevelId];
+  function mockUpgradeInstallation(UpgradeQueue calldata _upgradeQueue, uint40 _gltr) external {
+    // check owner
+    require(IERC721(s.realmDiamond).ownerOf(_upgradeQueue.parcelId) == _upgradeQueue.owner, "TestInstallationFacet: Not owner");
+    // check coordinates
     RealmDiamond realm = RealmDiamond(s.realmDiamond);
 
-    // Validation checks
+    realm.checkCoordinates(_upgradeQueue.parcelId, _upgradeQueue.coordinateX, _upgradeQueue.coordinateY, _upgradeQueue.installationId);
+
+    // check unique hash
     bytes32 uniqueHash = keccak256(
       abi.encodePacked(_upgradeQueue.parcelId, _upgradeQueue.coordinateX, _upgradeQueue.coordinateY, _upgradeQueue.installationId)
     );
-    require(s.upgradeHashes[uniqueHash] == 0, "InstallationUpgradeFacet: Upgrade hash not unique");
-    LibInstallation.checkUpgrade(_upgradeQueue, _gotchiId, realm);
 
-    // For easier testing, we min gltr instead of reverting
-    _gltr = nextInstallation.craftTime < _gltr ? nextInstallation.craftTime : _gltr;
+    //The same upgrade cannot be queued twice
+    require(s.upgradeHashes[uniqueHash] == 0, "TestInstallationFacet: Upgrade hash not unique");
 
+    s.upgradeHashes[uniqueHash] = _upgradeQueue.parcelId;
+
+    //current installation
+    InstallationType memory prevInstallation = s.installationTypes[_upgradeQueue.installationId];
+
+    //next level
+    InstallationType memory nextInstallation = s.installationTypes[prevInstallation.nextLevelId];
+
+    require(prevInstallation.nextLevelId > 0, "TestInstallationFacet: Maximum upgrade reached");
+    require(prevInstallation.installationType == nextInstallation.installationType, "TestInstallationFacet: Wrong installation type");
+    require(prevInstallation.alchemicaType == nextInstallation.alchemicaType, "TestInstallationFacet: Wrong alchemicaType");
+    require(prevInstallation.level == nextInstallation.level - 1, "TestInstallationFacet: Wrong installation level");
+
+    //prevent underflow if user sends too much GLTR
+    if (_gltr > nextInstallation.craftTime) revert("TestInstallationFacet: Too much GLTR");
+
+    //Confirm upgrade immediately
     if (nextInstallation.craftTime - _gltr == 0) {
-      //Confirm upgrade immediately
-      emit UpgradeTimeReduced(0, _upgradeQueue.parcelId, _upgradeQueue.coordinateX, _upgradeQueue.coordinateY, _gltr);
-      LibInstallation.upgradeInstallation(_upgradeQueue, nextLevelId, realm);
+      realm.upgradeInstallation(
+        _upgradeQueue.parcelId,
+        _upgradeQueue.installationId,
+        prevInstallation.nextLevelId,
+        _upgradeQueue.coordinateX,
+        _upgradeQueue.coordinateY
+      );
     } else {
-      s.upgradeHashes[uniqueHash] = _upgradeQueue.parcelId;
+      UpgradeQueue memory upgrade = UpgradeQueue(
+        _upgradeQueue.owner,
+        _upgradeQueue.coordinateX,
+        _upgradeQueue.coordinateY,
+        uint40(block.number) + nextInstallation.craftTime - _gltr,
+        false,
+        _upgradeQueue.parcelId,
+        _upgradeQueue.installationId
+      );
+      s.upgradeQueue.push(upgrade);
 
-      // Set the ready block and claimed flag before adding to the queue
-      _upgradeQueue.readyBlock = uint40(block.number) + nextInstallation.craftTime - _gltr;
-      _upgradeQueue.claimed = false;
-      LibInstallation.addToUpgradeQueue(_upgradeQueue, realm);
+      // update upgradeQueueLength
+      realm.addUpgradeQueueLength(_upgradeQueue.parcelId);
+
+      // Add to indexing helper to help for efficient getter
+      s.parcelIdToUpgradeIds[_upgradeQueue.parcelId].push(s.upgradeQueue.length - 1);
+
+      emit UpgradeInitiated(
+        _upgradeQueue.parcelId,
+        _upgradeQueue.coordinateX,
+        _upgradeQueue.coordinateY,
+        block.number,
+        uint40(block.number) + nextInstallation.craftTime - _gltr,
+        _upgradeQueue.installationId
+      );
+      emit UpgradeQueued(_upgradeQueue.owner, _upgradeQueue.parcelId, s.upgradeQueue.length - 1);
     }
   }
