--- scripts/realm/upgrades/upgrade-batchEquip.ts
@@ -1,49 +1,0 @@
-import { run, ethers, network } from "hardhat";
-import { varsForNetwork } from "../../../constants";
-import {
-  convertFacetAndSelectorsToString,
-  DeployUpgradeTaskArgs,
-  FacetsAndAddSelectors,
-} from "../../../tasks/deployUpgrade";
-
-export async function batchEquipUpgrade() {
-  const diamondUpgrader = "0xa370f2ADd2A9Fba8759147995d6A0641F8d7C119";
-
-  const c = await varsForNetwork(ethers);
-
-  const BatchEquip = `tuple(uint256[] types, bool[] equip, uint256[] ids, uint256[] x, uint256[] y)`;
-
-  const facets: FacetsAndAddSelectors[] = [
-    {
-      facetName: "RealmFacet",
-      addSelectors: [
-        `function batchEquip(uint256 _realmId,uint256 _gotchiId,${BatchEquip} memory _equipParams ,bytes[] memory _signatures) external`,
-      ],
-      removeSelectors: [],
-    },
-  ];
-
-  const joined = convertFacetAndSelectorsToString(facets);
-
-  const args: DeployUpgradeTaskArgs = {
-    diamondUpgrader: diamondUpgrader,
-    diamondAddress: c.realmDiamond,
-    facetsAndAddSelectors: joined,
-    useLedger: true,
-    useMultisig: false,
-    initAddress: ethers.constants.AddressZero,
-    initCalldata: "0x",
-  };
-
-  await run("deployUpgrade", args);
-}
-
-if (require.main === module) {
-  batchEquipUpgrade()
-    .then(() => process.exit(0))
-    // .then(() => console.log('upgrade completed') /* process.exit(0) */)
-    .catch((error) => {
-      console.error(error);
-      process.exit(1);
-    });
-}
