--- scripts/realm/upgrades/upgrade-checkStartCoordinate.ts
@@ -1,60 +1,0 @@
-import { ethers, network, run } from "hardhat";
-import { varsForNetwork } from "../../../constants";
-import {
-  convertFacetAndSelectorsToString,
-  DeployUpgradeTaskArgs,
-  FacetsAndAddSelectors,
-} from "../../../tasks/deployUpgrade";
-import { RealmGridFacet, RealmFacet } from "../../../typechain";
-
-export async function upgradeRealm() {
-  const c = await varsForNetwork(ethers);
-
-  const diamondUpgrader = "0xa370f2ADd2A9Fba8759147995d6A0641F8d7C119";
-
-  const realmFacets: FacetsAndAddSelectors[] = [
-    {
-      facetName: "RealmGettersAndSettersFacet",
-      addSelectors: [
-        "function verifyAccessRight(uint256 _realmId,uint256 _gotchiId,uint256 _actionRight,address _sender) external view",
-      ],
-      removeSelectors: [],
-    },
-    {
-      facetName: "RealmFacet",
-      addSelectors: [],
-      removeSelectors: [],
-    },
-    {
-      facetName: "AlchemicaFacet",
-      addSelectors: [],
-      removeSelectors: [],
-    },
-  ];
-
-  const realmJoined = convertFacetAndSelectorsToString(realmFacets);
-
-  console.log("realm diamond:", c.realmDiamond);
-
-  const realmArgs: DeployUpgradeTaskArgs = {
-    diamondUpgrader: diamondUpgrader,
-    diamondAddress: c.realmDiamond,
-    facetsAndAddSelectors: realmJoined,
-    useLedger: true,
-    useMultisig: false,
-    initCalldata: "0x",
-    initAddress: ethers.constants.AddressZero,
-  };
-
-  await run("deployUpgrade", realmArgs);
-}
-
-if (require.main === module) {
-  upgradeRealm()
-    .then(() => process.exit(0))
-    // .then(() => console.log('upgrade completed') /* process.exit(0) */)
-    .catch((error) => {
-      console.error(error);
-      process.exit(1);
-    });
-}
